Edited the transfer function to reflect the fact that the padlock and donation bar display settings have been moved to the Style Settings tab

+++ includes/settings/functions.php
@@ -305,7 +305,7 @@ function coil_messaging_settings_validation( $messaging_settings ) : array {
 /**
  * Allow the radio buttons that select the padlock and donation bar
  * display checkboxes to be properly validated
- * 
+ *
  * @param array $style_settings The padlock and donation bar display checkboxes
  *
  * @return array
@@ -971,17 +971,13 @@ function dismiss_welcome_notice() {
 /**
  * Translate customizer settings
  *
- * If a user has settings which they saved in the customizer, switch them to settings saved in the wp_options table
+ * If a user has message settings which they saved in the customizer, switch them to settings saved in the wp_options table
  *
  */
 
 function transfer_customizer_message_settings() {
 
 	$existing_options = get_option( 'coil_messaging_settings_group' );
-
-	if ( ! empty( $existing_options ) ) {
-		return;
-	}
 
 	$messaging_settings = [];
 
@@ -993,12 +989,26 @@ function transfer_customizer_message_settings() {
 	$coil_learn_more_button_text     = 'coil_learn_more_button_text';
 	$coil_learn_more_button_link     = 'coil_learn_more_button_link';
 
+	$customizer_empty = (
+		get_theme_mod( $coil_partial_gating_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_unsupported_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_verifying_status_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_unable_to_verify_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_voluntary_donation_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_learn_more_button_text, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_learn_more_button_link, 'null' ) !== 'null'
+	);
+
+	if ( $customizer_empty ) {
+		return;
+	}
+
 	// Using 'null' for comparrison becasue custom messages that were deleted remain in the database with the value false, but still need to be removed.
 	if ( get_theme_mod( $coil_partial_gating_message, 'null' ) !== 'null' ) {
 		$messaging_settings['coil_partially_gated_content_message'] = get_theme_mod( $coil_partial_gating_message );
 		remove_theme_mod( $coil_partial_gating_message );
 	}
-	if ( get_theme_mod( $coil_unsupported_message , 'null' ) !== 'null' ) {
+	if ( get_theme_mod( $coil_unsupported_message, 'null' ) !== 'null' ) {
 		$messaging_settings['coil_fully_gated_content_message'] = get_theme_mod( $coil_unsupported_message );
 		remove_theme_mod( $coil_unsupported_message );
 	}
@@ -1028,9 +1038,21 @@ function transfer_customizer_message_settings() {
 		remove_theme_mod( $coil_learn_more_button_link );
 	}
 
-	update_option( 'coil_messaging_settings_group', $messaging_settings );
+	if ( false !== $existing_options ) {
+		update_option( 'coil_messaging_settings_group', array_merge( $existing_options, $messaging_settings ) );
+	} else {
+		update_option( 'coil_messaging_settings_group', $messaging_settings );
+	}
 
 }
+
+/**
+ * Translate customizer settings
+ *
+ * If a user has gating settings which they saved in the settings panel,
+ * switch them to settings saved in the wp_options table under coil_monetization_settings_group
+ *
+ */
 
 function transfer_customizer_monetization_settings() {
 
@@ -1040,23 +1062,14 @@ function transfer_customizer_monetization_settings() {
 
 	// If the setting has already been saved or transferred then simply return
 	// Using 'null' for comparrison becasue if the padlock and support creator messages were unselected they were stored in the database with the value false, but still need to be transferred.
-	if ( get_option( 'coil_monetization_settings_group' ) || ( ! get_option( 'coil_content_settings_posts_group' ) && 'null' === get_theme_mod( 'coil_title_padlock', 'null' ) && 'null' === get_theme_mod( 'coil_show_donation_bar', 'null' ) ) ) {
+	if ( ! get_option( 'coil_content_settings_posts_group' ) ) {
 		return;
 	}
 
-	$coil_title_padlock     = 'coil_title_padlock';
-	$coil_show_donation_bar = 'coil_show_donation_bar';
-
-	$new_monetization_settings = [
-		'coil_title_padlock'     => get_theme_mod( $coil_title_padlock, true ),
-		'coil_show_donation_bar' => get_theme_mod( $coil_show_donation_bar, true ),
-	];
-
-	remove_theme_mod( $coil_title_padlock );
-	remove_theme_mod( $coil_show_donation_bar );
+	$new_monetization_settings = [];
 
 	// Before moving post gating settings across check that post types and gating types are valid.
-	$supported_post_types = Coil\get_supported_post_types( 'names' );
+	$supported_post_types     = Coil\get_supported_post_types( 'names' );
 	$supported_gating_options = array_keys( Gating\get_monetization_setting_types() );
 
 	if ( ! empty( $supported_post_types ) && ! empty( $previous_gating_options ) ) {
@@ -1073,5 +1086,44 @@ function transfer_customizer_monetization_settings() {
 		update_option( 'coil_monetization_settings_group', array_merge( $existing_options, $new_monetization_settings ) );
 	} else {
 		update_option( 'coil_monetization_settings_group', $new_monetization_settings );
+	}
+}
+
+/**
+ * Translate customizer settings
+ *
+ * If a user has style settings which they saved in the customizer, switch them to settings saved in the wp_options table
+ *
+ */
+
+function transfer_customizer_style_settings() {
+
+	$existing_options = get_option( 'coil_style_settings_group' );
+
+	// If the setting has already been saved or transferred then simply return
+	// Using 'null' for comparrison becasue if the padlock and support creator messages were unselected they were stored in the database with the value false, but still need to be transferred.
+	if ( 'null' === get_theme_mod( 'coil_title_padlock', 'null' ) && 'null' === get_theme_mod( 'coil_show_donation_bar', 'null' ) ) {
+		return;
+	}
+
+	$coil_title_padlock     = 'coil_title_padlock';
+	$coil_show_donation_bar = 'coil_show_donation_bar';
+
+	$new_style_settings = [];
+
+	if ( get_theme_mod( $coil_title_padlock, 'null' ) !== 'null' ) {
+		$new_style_settings['coil_title_padlock'] = get_theme_mod( $coil_title_padlock, true );
+		remove_theme_mod( $coil_title_padlock );
+	}
+
+	if ( get_theme_mod( $coil_show_donation_bar, 'null' ) !== 'null' ) {
+		$new_style_settings['coil_show_donation_bar'] = get_theme_mod( $coil_show_donation_bar, true );
+		remove_theme_mod( $coil_show_donation_bar );
+	}
+
+	if ( false !== $existing_options ) {
+		update_option( 'coil_style_settings_group', array_merge( $existing_options, $new_style_settings ) );
+	} else {
+		update_option( 'coil_style_settings_group', $new_style_settings );
 	}
 }
